[#1551] Fix doc links and clang-tidy warnings

+++ iceoryx2-cxx/include/iox2/internal/callback_context.hpp
@@ -67,14 +67,13 @@ inline auto ctx_cast(void* ptr) -> CallbackContext<T>* {
     return static_cast<CallbackContext<T>*>(ptr);
 }
 
-#include <iostream>
-
-inline auto override_callback(size_t value, iox2_callback_context ctx) -> uint64_t {
+inline auto override_callback(size_t value, iox2_callback_context ctx) -> size_t {
     // NOLINTNEXTLINE(cppcoreguidelines-owning-memory) must be raw pointer because of C interface
     auto* context = ctx_cast<iox2::bb::StaticFunction<size_t(size_t)>*>(ctx);
     auto ret_val = (*context->value())(value);
     // NOLINTNEXTLINE(cppcoreguidelines-owning-memory) must be raw pointer because of C interface
     delete context->value();
+    // NOLINTNEXTLINE(cppcoreguidelines-owning-memory) must be raw pointer because of C interface
     delete context;
     return ret_val;
 }
